feat: Add prompt utilities for input and select prompts
